pyproject.toml
[project]
name = "dydx-v4-grpc-streaming-client"
version = "0.2.0"
description = "dYdX V4 grpc full node streaming client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.67.0",
    "grpcio-tools==1.64.1",
    "protobuf==5.28.1",
    "pyyaml==6.0.1",
    "requests~=2.32.2",
    "ruff>=0.9.9",
    "sortedcontainers==2.4.0",
    "v4-proto==6.0.8",
    "websockets==12.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "types-pyyaml~=6.0.12.20241230",
    "types-requests~=2.32.0.20250301",
]

[tool.uv]
required-version = ">=0.5.0"

[tool.ruff]
line-length = 120
exclude = [
    ".mypy_cache",
    ".venv",
    "venv",
]

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "v4_proto.amino.*,v4_proto.dydxprotocol.*,v4_proto.cosmos_proto.*,v4_proto.gogoproto.*,v4_proto.google.*"
disable_error_code = "misc"

[[tool.mypy.overrides]]
module = "v4_proto.dydxprotocol.clob.*"
disable_error_code = "import-untyped"
